reworks activation
