+++ src/components/entity-hoverable-description/entity-description-hover.module.css
